18/6/21

+++ app/src/main/java/com/example/restaurant_app/modelmanager/getOrderItem/Order.java
@@ -1,4 +1,3 @@
-
 package com.example.restaurant_app.modelmanager.getOrderItem;
 
 import com.google.gson.annotations.Expose;
@@ -16,7 +15,7 @@ public class Order {
     private String paymentStatus;
     @SerializedName("grandTotal")
     @Expose
-    private Integer grandTotal;
+    private Double grandTotal;
     @SerializedName("paymentMethod")
     @Expose
     private String paymentMethod;
@@ -67,11 +66,11 @@ public class Order {
         this.paymentStatus = paymentStatus;
     }
 
-    public Integer getGrandTotal() {
+    public Double getGrandTotal() {
         return grandTotal;
     }
 
-    public void setGrandTotal(Integer grandTotal) {
+    public void setGrandTotal(Double grandTotal) {
         this.grandTotal = grandTotal;
     }
